% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_compare_plot.R
\name{create_ge_compare_plot_report_fbk}
\alias{create_ge_compare_plot_report_fbk}
\title{Comparison Plot Creator Function For FBK}
\usage{
create_ge_compare_plot_report_fbk(pn_cur_ge_label, pn_prev_ge_label,
  pl_plot_opts = NULL, pb_debug = FALSE)
}
\arguments{
\item{pl_plot_opts}{list of options specifying input for plot report creator}
}
\description{
Comparison reports containing all generated plots of a GE side-by-side
with the plots of the previous GE are constructed for a given trait
group.
}
\examples{
\dontrun{
create_ge_compare_plot_report_fbk(pn_cur_ge_label  = 1908,
                                  pn_prev_ge_label = 1904)
}

}
