% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{generic_logger}
\alias{generic_logger}
\title{Generic Logging Function}
\usage{
generic_logger(ps_level, ps_caller, ps_msg)
}
\description{
This function provides the generic loggin mechanism
It is mostly used through level-specific wrapper functions.
}
