% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_plh}
\alias{replace_plh}
\title{Substitution Of Text-Markers With Replacement Values In Report Text}
\usage{
replace_plh(ps_report_text, pl_replacement, pb_debug = FALSE)
}
\description{
Given a report text with placeholders and a list of placeholderpattern-
replacement-value pairs, the function loops through the list and replaces the
placeholders by the replacement values. The single elements of the replacement
list must be a list again with two elements having names 'pattern' and
'replacement' where the former contains the placeholder pattern and the latter
contains the replacement value.
}
