% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwsroutinetools.R
\docType{package}
\name{zwsroutinetools}
\alias{zwsroutinetools}
\alias{zwsroutinetools-package}
\title{zwsroutinetools: A package providing tools for a Genetic Evaluation (GE) Routine}
\description{
The package `zwsroutinetools` provides R-functions and shell-scripts that are
considered to be useful. The set of provided tools can be grouped into different
focus categories. So far just one category of generating reports of comparison
with previous GE runs.
}
